fix: type rename and add tsx
