+++ package.json
@@ -63,6 +63,7 @@
     "npm-run-all": "4.1.5",
     "prettier": "3.3.2",
     "ts-jest": "29.1.5",
+    "tsx": "4.15.6",
     "typescript": "5.4.5",
     "typescript-eslint": "7.13.1",
     "wrangler": "3.60.3"
